requirements.txt
praw==7.8.1
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.8.5